requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
alembic
python-multipart
jinja2
playwright
easyocr
opencv-python
numpy
chartjs
pydantic
pydantic-settings
xhtml2pdf
pyjwt
supabase
python-dotenv
